\name{pool}
\alias{pool}
\docType{data}
\title{
Seed data used in our article [TODO ref]
}
\description{
30000 gene expression levels sampled randomly from an Illumina WG-6 chip.
}
\usage{data("pool")}
\format{
  A vector of 30000 log2 gene expression values
}
\references{
[TODO ref]
}
\examples{
data(pool)
hist(pool)
}
\keyword{datasets}
